+++ app/src/main/java/com/github/nakawai/newsreader/KoinInjector.kt
@@ -2,9 +2,9 @@ package com.github.nakawai.newsreader
 
 import com.github.nakawai.newsreader.data.Repository
 import com.github.nakawai.newsreader.data.RepositoryImpl
-import com.github.nakawai.newsreader.data.db.NYTimesLocalDataSource
+import com.github.nakawai.newsreader.domain.datasource.NYTimesLocalDataSource
 import com.github.nakawai.newsreader.data.db.NYTimesLocalDataSourceImpl
-import com.github.nakawai.newsreader.data.network.NYTimesRemoteDataSource
+import com.github.nakawai.newsreader.domain.datasource.NYTimesRemoteDataSource
 import com.github.nakawai.newsreader.data.network.NYTimesRemoteDataSourceImpl
 import com.github.nakawai.newsreader.presentation.articles.ArticlesViewModel
 import com.github.nakawai.newsreader.presentation.details.DetailsViewModel
