+++ app/src/main/java/com/github/nakawai/newsreader/domain/datasource/NYTimesRemoteDataSource.kt
@@ -1,7 +1,7 @@
-package com.github.nakawai.newsreader.data.network
+package com.github.nakawai.newsreader.domain.datasource
 
 import com.github.nakawai.newsreader.data.network.response.StoryResponseItem
-import com.github.nakawai.newsreader.domain.story.Section
+import com.github.nakawai.newsreader.domain.entities.Section
 
 interface NYTimesRemoteDataSource {
     suspend fun fetchData(section: Section): List<StoryResponseItem>
